requirements.txt
Django==1.9.6
django-crispy-forms==1.6.0
django-filter==0.13.0
django-guardian==1.4.4
djangorestframework==3.3.3
Markdown==2.6.6
six==1.10.0
